% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalBgSpatialKld}
\alias{CalBgSpatialKld}
\title{Compute Background 2D Kernel Density}
\usage{
CalBgSpatialKld(coords, gx, gy, h)
}
\arguments{
\item{coords}{coordinate matrix.}

\item{gx}{Vector grid points in x direction, see(\code{seq(lims[1], lims[2], length.out=200)}).}

\item{gy}{Vector grid points in y direction, see(\code{seq(lims[3], lims[4], length.out=200)}).}

\item{h}{The vector of bandwidths for x and y directions, defaults to normal reference bandwidth
(see MASS::bandwidth.nrd), A scalar value will be taken to apply to both directions (see ks::hpi).}
}
\description{
Compute Background 2D Kernel Density
}
