% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalSpatialKldCpp}
\alias{CalSpatialKldCpp}
\title{Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted And Statistical Test With Permutation.}
\usage{
CalSpatialKldCpp(coords, d, l, h, n = 100L, random_times = 200L, seed = 123L)
}
\arguments{
\item{coords}{coordinate matrix.}

\item{d}{matrix (the expression of gene or score of pathway).}

\item{l}{The limits of the rectangle covered by the grid as c(xl, xu, yl, yu).}

\item{h}{The vector of bandwidths for x and y directions, defaults to normal reference bandwidth
(see bandwidth.nrd), A scalar value will be taken to apply to both directions (see ks::hpi).}

\item{n}{the Number of grid points in each direction, default is 100.}

\item{random_times}{the permutation numbers for each weight to test whether
it is significantly, default is 200.}

\item{seed}{The random seed to use to evaluate, default 123.}
}
\description{
Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted And Statistical Test With Permutation.
}
