% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cluster.assign.R
\name{cluster.assign}
\alias{cluster.assign}
\alias{cluster.assign,SingleCellExperiment-method}
\alias{cluster.assign,SingleCellExperiment}
\alias{cluster.assign,SVPExperiment-method}
\alias{cluster.assign,SVPExperiment}
\title{clusting and assign the label for each feature(especifily the gene sets).}
\usage{
cluster.assign(
  data,
  assay.type = "affi.score",
  ncluster = 2,
  random.seed = 1024,
  BPPARAM = SerialParam(),
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{cluster.assign}{SingleCellExperiment}(
  data,
  assay.type = "affi.score",
  ncluster = 2,
  random.seed = 1024,
  BPPARAM = SerialParam(),
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{cluster.assign}{SVPExperiment}(
  data,
  assay.type = "affi.score",
  ncluster = 2,
  random.seed = 1024,
  BPPARAM = SerialParam(),
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)
}
\arguments{
\item{data}{A \linkS4class{SVPExperiment}, which has run \code{sc.rwr} or \code{detect.svp}, or
a \linkS4class{SingleCellExperiment} which was extracted from \linkS4class{SVPExperiment} using
\code{gsvaExp} function.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{affi.score}.}

\item{ncluster}{integer the number cluster for each feature, see details, default is 2.}

\item{random.seed}{integer random number to be reproducted.}

\item{BPPARAM}{A BiocParallelParam object specifying whether the built of KNN should be parallelized
default is \code{SerialParam()}, meaning no parallel. You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)}
to parallel it, the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{...}{dot parameters}
}
\value{
if input is a \linkS4class{SVPExperiment}, output will be also a \linkS4class{SVPExperiment}, and the result assay
was stored in assay of the specified \code{gsvaexp}, which is a \linkS4class{SingleCellExperiment}. If input is a
\linkS4class{SingleCellExperiment} (which is extracted from \linkS4class{SVPExperiment} using \code{gsvaExp()} funtion),
output will be a \linkS4class{SingleCellExperiment}, the result can be extracted using \code{assay()} function.
}
\description{
clusting and assign the label for each feature(especifily the gene sets).
}
\details{
\code{cluster.assign} using \code{kmeans} to classfing the cells with the each activity score of the gene sets. When the argument
\code{ncluster = 2}, meaning the gene sets or pathway was activated in the \code{1} labeled cells. When the argument \code{ncluster}
larger than 2, the larger the cluster label, the stronger the activation of the cell in the specified pathway.
}
\seealso{
to calculate the activity score of gene sets or pathway: \code{\link{sc.rwr}}.
}
