% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pdac_a_sce}
\alias{data_pdac_a_sce}
\alias{pdac_a_sce}
\title{the single cell gene profiler of a PDAC sample}
\format{
S4 class:SingleCellExperiment
}
\value{
a \linkS4class{SingleCellExperiment} object
}
\description{
The single cell transcriptome of PDAC A sample from (doi:10.1038/s41587-019-0392-8)
}
\examples{
data(pdac_a_sce)
}
\keyword{data}
