% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pdac_a_spe}
\alias{data_pdac_a_spe}
\alias{pdac_a_spe}
\title{the spatial transcriptome of a PDAC sample}
\format{
S4 class:SpatialExperiment
}
\value{
a \linkS4class{SpatialExperiment} object
}
\description{
The spatial transcriptome of PDAC A sample from (doi:10.1038/s41587-019-0392-8)
}
\examples{
data(pdac_a_spe)
}
\keyword{data}
