% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sceSubPbmc}
\alias{data_sceSubPbmc}
\alias{sceSubPbmc}
\title{a subset data of pbmck3 from SeuratData}
\description{
a small SingleCellExperiment data set from pbmck3 which
contains 1304 genes and 800 cells (extract randomly)
}
\examples{
data(sceSubPbmc)
}
\keyword{data}
