% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SenMayo}
\alias{data_SenMayo}
\alias{SenMayoSymbol}
\title{A gene set identifies senescent cells and predicts senescence-associated
pathways across tissues}
\format{
list
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41467-022-32552-1/MediaObjects/41467_2022_32552_MOESM4_ESM.xlsx}
}
\value{
a list object
}
\description{
SenMayoSymbol is a gene symbol list that can be used to identify senescent cells and predicts
senescence-associated pathways across tissues
}
\examples{
data(SenMayoSymbol)
}
\references{
Saul, D., Kosinsky, R.L., Atkinson, E.J. et al. A new gene set identifies senescent
cells and predicts senescence-associated pathways across tissues. Nat Commun 13, 4827 (2022).
https://doi.org/10.1038/s41467-022-32552-1
}
\keyword{data}
