% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_color.R
\name{extract_color}
\alias{extract_color}
\title{extract image color feature}
\usage{
extract_color(img, beta = 2, coords, alpha = 1)
}
\arguments{
\item{img}{raster object}

\item{beta}{the size of region to merge}

\item{coords}{the coordinate to spots to extract color}

\item{alpha}{the color scale}
}
\description{
extract image color feature
}
