% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExtractFeatureScoreCpp}
\alias{ExtractFeatureScoreCpp}
\title{Extract the score of gene in each gene sets}
\usage{
ExtractFeatureScoreCpp(x, rnm, cnm, g)
}
\arguments{
\item{x}{the score sparse matrix of gene for each gene sets.}

\item{rnm}{the row names of x matrix.}

\item{cnm}{the col names of x matrix.}

\item{g}{a list of gene set.}
}
\description{
Extract the score of gene in each gene sets
}
