% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPDist}
\alias{fastPDist}
\title{Obtain the pair distance of row between \code{A} and \code{B} matrix}
\usage{
fastPDist(A, B)
}
\arguments{
\item{A}{matrix which number of column should be equal to column number of \code{B}.}

\item{B}{matrix which number of column should be equal to column number of \code{A}.}
}
\value{
a distance matrix of row feature in \code{A} and \code{B}.
}
\description{
Obtain the pair distance of row between \code{A} and \code{B} matrix
}
