% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findIntervalCpp}
\alias{findIntervalCpp}
\title{Find Interval Numbers or Indices in C++}
\usage{
findIntervalCpp(x, breaks)
}
\arguments{
\item{x}{numeric vector (orignial).}

\item{breaks}{numeric vector (new).}
}
\description{
Find Interval Numbers or Indices in C++
}
