% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svg.R
\name{kldSVG}
\alias{kldSVG}
\alias{kldSVG,SingleCellExperiment-method}
\alias{kldSVG,SingleCellExperiment}
\alias{kldSVG,SVPExperiment-method}
\alias{kldSVG,SVPExperiment}
\title{Detecting the spatially or single cell variable features with Kullback–Leibler
divergence of 2D Weighted Kernel Density}
\usage{
kldSVG(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "fdr",
  sv.BPPARAM = SerialParam(),
  cells = NULL,
  features = NULL,
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{kldSVG}{SingleCellExperiment}(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "fdr",
  sv.BPPARAM = SerialParam(),
  cells = NULL,
  features = NULL,
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{kldSVG}{SVPExperiment}(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "fdr",
  sv.BPPARAM = SerialParam(),
  cells = NULL,
  features = NULL,
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object with contains \code{UMAP} or \code{TSNE},
or a \linkS4class{SpatialExperiment} object, or a \linkS4class{SVPExperiment} object with specified
\code{gsvaexp} argument.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{logcounts}.}

\item{sv.used.reduction}{character used as spatial coordinates to detect SVG, default is \code{UMAP},
if \code{data} has \code{spatialCoords}, which will be used as spatial coordinates.}

\item{sv.grid.n}{numeric number of grid points in the two directions to estimate 2D weighted kernel
density, default is 100.}

\item{sv.permutation}{numeric the number of permutation for each single feature to detect the
signicantly spatially or single cell variable features, default is 100.}

\item{sv.p.adjust.method}{character the method to adjust the pvalue of the result, default is \code{fdr}.}

\item{sv.BPPARAM}{A BiocParallelParam object specifying whether the identification of SV features should be parallelized
default is \code{SerialParam()}, meaning no parallel. You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)}
to parallel it, the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}. default is \code{SerialParam()}.}

\item{cells}{Vector specifying the subset of cells to be used for the calculation of the activaty score or identification
of SV features. This can be a character vector of cell names, an integer vector of column indices or a logical vector,
default is NULL, meaning all cells to be used for the calculation of the activaty score or identification of SV features.}

\item{features}{Vector specifying the subset of features to be used for the calculation of the activaty score or identification
of SV features. This can be a character vector of features names, an integer vector of row indices or a logical vector,
default is NULL, meaning all features to be used for the calculation of the activaty score or identification of SV features.}

\item{verbose}{logical whether print the intermediate message when running the program, default is TRUE.}

\item{random.seed}{numeric random seed number to repeatability, default is 1024.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{...}{additional parameters}
}
\description{
Detecting the spatially or single cell variable features with Kullback–Leibler
divergence of 2D Weighted Kernel Density
}
