% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mob_marker_genes}
\alias{mob_marker_genes}
\title{the marker genes of mouse olfactory bulb}
\format{
list
}
\value{
a list object with name
}
\description{
this is extracted from the single cell transcriptome of a mouse olfactory bulb from (doi:10.1016/j.celrep.2018.11.034)
}
\examples{
data(mob_marker_genes)
}
\keyword{data}
