% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mob_sce}
\alias{mob_sce}
\title{the single cell gene profiler of a mouse olfactory bulb}
\format{
S4 class:SingleCellExperiment
}
\value{
a \linkS4class{SingleCellExperiment} object
}
\description{
The single cell transcriptome of WT sample of mouse olfactory bulb from (doi:10.1016/j.celrep.2018.11.034)
}
\examples{
data(mob_sce)
}
\keyword{data}
