% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outergrid}
\alias{outergrid}
\title{Obtaion the difference between the grid points and original points}
\usage{
outergrid(grid, x)
}
\arguments{
\item{grid}{the grid points in one direction}

\item{x}{the original points in one direction}
}
\description{
Obtaion the difference between the grid points and original points
}
