% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelCalRWR}
\alias{parallelCalRWR}
\title{Computer the affinity score of all nodes in a graph to a seeds
using Random Walk with Restart}
\usage{
parallelCalRWR(x, v, restart = 0.75, stop_delta = 1e-10, stop_step = 50L)
}
\arguments{
\item{x}{a adjacency matrix of a graph.}

\item{v}{a matrix define sets of starting seeds, each column
corresponds to one set of seeds that a walker starts.}

\item{restart}{the restart probability used for RWR, it must be
between 0 and 1, default is .75.}

\item{stop_delta}{minimum threshold to stop RWR, default is 1e-10.}

\item{stop_step}{step number to stop RWR, default is 50.}
}
\value{
a matrix of affinity score of all nodes in a graph
}
\description{
Computer the affinity score of all nodes in a graph to a seeds
using Random Walk with Restart
}
