% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pred.cell.signature.R
\name{pred.feature.mode}
\alias{pred.feature.mode}
\alias{pred.feature.mode,SingleCellExperiment-method}
\alias{pred.feature.mode,SingleCellExperiment}
\alias{pred.feature.mode,SVPExperiment-method}
\alias{pred.feature.mode,SVPExperiment}
\title{predict the expression or activity score mode of the
gene sets or pathway activity score or other features.}
\usage{
pred.feature.mode(
  data,
  assay.type = "rwr.score",
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  mod0 = 2,
  BPPARAM = SerialParam(),
  name = "ModeFlag",
  features = NULL,
  cells = NULL,
  ...
)

\S4method{pred.feature.mode}{SingleCellExperiment}(
  data,
  assay.type = "rwr.score",
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  mod0 = 2,
  BPPARAM = SerialParam(),
  name = "ModeFlag",
  features = NULL,
  cells = NULL,
  ...
)

\S4method{pred.feature.mode}{SVPExperiment}(
  data,
  assay.type = "rwr.score",
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  mod0 = 2,
  BPPARAM = SerialParam(),
  name = "ModeFlag",
  features = NULL,
  cells = NULL,
  ...
)
}
\arguments{
\item{data}{A \linkS4class{SVPExperiment}, which has run \code{runSGSA} or \code{detect.svp}, or
a \linkS4class{SingleCellExperiment} which was extracted from \linkS4class{SVPExperiment} using
\code{gsvaExp} function.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{rwr.score}.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{mod0}{integer Number of modes for which the critical bandwidth is calculated, default is 2.}

\item{BPPARAM}{A BiocParallelParam object specifying whether perform the analysis parallelly using
\code{BiocParallel} default is \code{SerialParam()}, meaning no parallel.
You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)} to parallel it,
the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}. default is \code{SerialParam()}.}

\item{name}{character the assay name in \code{assay} of the result, default is \code{ModeFlag}.}

\item{features}{Vector specifying the subset of features to be used for the analysis, if \code{gsvaexp} is
specified, it should be the rownames or rownames index of \code{gsvaexp(data)}, default is NULL, meaning
all features will be analyzed.}

\item{cells}{Vector specifying the subset of cells to be used for the analysis, default is NULL, meaning
all cells will be analyzed.}

\item{...}{dot parameters.}
}
\value{
if input is a \linkS4class{SVPExperiment}, output will be also a \linkS4class{SVPExperiment}, and the result
was stored in the \code{assays} of \linkS4class{SingleCellExperiment} of the specified \code{gsvaexp},
which is a \linkS4class{SingleCellExperiment}.  If input is a \linkS4class{SingleCellExperiment} (which is extracted from
\linkS4class{SVPExperiment} using \code{gsvaExp()} funtion), output will be a \linkS4class{SingleCellExperiment},
the result can be extracted using \code{assay()} function with specified name in default is \code{ModeFlag}.
}
\description{
this function will return a certain number of modes for each features using the estimation of
the location of modes and antimodes with the expression or activity score.
}
\examples{
library(SpatialExperiment)
# This example data was extracted from the
# result of runSGSA with gsvaExp function.
data(hpda_spe_cell_dec)
assays(hpda_spe_cell_dec)
hpda_spe_cell_dec <- hpda_spe_cell_dec |> 
   pred.feature.mode(assay.type = 'rwr.score', 
     mod0 = 2, 
     BPPARAM=BiocParallel::MulticoreParam(workers=2,progressbar=TRUE)
   )
# The result will be saved in assays of hpda_spe_cell_dec
# and it can be extracted using assay function.
assays(hpda_spe_cell_dec)
assay(hpda_spe_cell_dec, 'ModeFlag') |> t() |> head()
# We extract the activity mode of Cancer clone A and Cancer clone B, then visualize them
assay(hpda_spe_cell_dec, 'ModeFlag') |> t() |> 
 magrittr::extract(,c(2, 3)) |> 
 as.matrix() |> 
 data.frame() |> 
 dplyr::mutate_all(as.factor) |> 
 DataFrame() -> colData(hpda_spe_cell_dec)
\dontrun{
  library(ggplot2)
  library(ggsc)
  p1 <- sc_spatial(hpda_spe_cell_dec, 
             features = rownames(hpda_spe_cell_dec), 
             mapping = aes(x=x,y=y, color=Cancer.clone.A), 
             plot.pie = T, 
             pie.radius.scale = .8, 
             bg_circle_radius = 1.1, 
             color=NA, 
             linewidth=2
  ) + 
  scale_color_manual(values=c('white', 'black'))
  p1
  f1 <- sc_spatial(hpda_spe_cell_dec, features="Cancer clone A", 
             mapping=aes(x=x,y=y),
             pointsize=10
  ) + 
  geom_scattermore2(
    mapping = aes(bg_color=Cancer.clone.A, subset=Cancer.clone.A==1), 
    bg_line_width = .15, 
    pointsize = 8
  ) + 
  scale_bg_color_manual(values=c('black')) 
  f1
  f2 <- sc_spatial(hpda_spe_cell_dec, features="Cancer clone B",
             mapping=aes(x=x,y=y),
             pointsize=10
  ) +
  geom_scattermore2(
    mapping = aes(bg_color=Cancer.clone.B, subset=Cancer.clone.B==1),
    bg_line_width = .15,
    pointsize = 8
  ) +
  scale_bg_color_manual(values=c('black')) 
  f2  
}
}
\seealso{
to calculate the activity score of gene sets or pathway: \code{\link{runSGSA}},
to keep the max gene set or pathway activity score of cell: \code{\link{cluster.assign}}.
}
\author{
Shuangbin Xu
}
