% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{imgData}
\alias{imgData<-}
\alias{spatialCoordsNames}
\alias{spatialCoordsNames<-}
\title{Objects exported from other packages}
\value{
function
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SpatialExperiment}{\code{\link[SpatialExperiment:imgData]{imgData()}}, \code{\link[SpatialExperiment:imgData<-]{imgData<-()}}, \code{\link[SpatialExperiment:spatialCoords]{spatialCoords()}}, \code{\link[SpatialExperiment:spatialCoords<-]{spatialCoords<-()}}, \code{\link[SpatialExperiment:spatialCoordsNames]{spatialCoordsNames()}}, \code{\link[SpatialExperiment:spatialCoordsNames<-]{spatialCoordsNames<-()}}}
}}

