% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-mca.R
\name{runMCA}
\alias{runMCA}
\alias{runMCA,SingleCellExperiment-method}
\alias{runMCA,SingleCellExperiment}
\title{Run Multiple Correspondence Analysis}
\usage{
runMCA(
  data,
  assay.type = "logcounts",
  reduction.name = "MCA",
  ncomponents = 30,
  subset.row = NULL,
  consider.spcoord = TRUE,
  ...
)

\S4method{runMCA}{SingleCellExperiment}(
  data,
  assay.type = "logcounts",
  reduction.name = "MCA",
  ncomponents = 50,
  subset.row = NULL,
  consider.spcoord = TRUE,
  ...
)
}
\arguments{
\item{data}{a SingleCellExperiment object}

\item{assay.type}{which expressed data to be pulled to run,
default is \code{logcounts}.}

\item{reduction.name}{name of the reduction result, default is \code{MCA}.}

\item{ncomponents}{number of components to compute and store, default is 30.}

\item{subset.row}{Vector specifying the subset of features to be used for
dimensionality reduction. This can be a character vector of row names,
an integer vector of row indices or a logical vector, default is NULL, meaning
all features to be used for dimensionality reduction.}

\item{consider.spcoord}{whether consider the spatial coords to run MCA with
the features of data, default is TRUE.}

\item{...}{additional parameters, meaningless now.}
}
\description{
Perform a Multiple Correspondence Analysis (MCA) on cells, based on
the expression data in a SingleCellExperiment object. It is modified based on the
\code{RunMCA} of \code{CelliD} with the source codes of C++.
}
