% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sc.rwr.R
\name{sc.rwr}
\alias{sc.rwr}
\alias{sc.rwr,SingleCellExperiment-method}
\alias{sc.rwr,SingleCellExperiment}
\title{Calculate the activity of gene sets in spatial or single-cell data with Restart Walk with Restart}
\usage{
sc.rwr(
  data,
  gset.idx.list,
  gsvaExp.name = "gset1.rwr",
  min.sz = 10,
  max.sz = Inf,
  gene.occurrence.rate = 0.4,
  assay.type = "logcounts",
  knn.mca.consider.spcoord = TRUE,
  knn.used.reduction.dims = 30,
  knn.BPPARAM = SerialParam(),
  knn.graph.weighted = TRUE,
  knn.graph.directed = FALSE,
  knn.k.use = 500,
  rwr.restart = 0.7,
  rwr.normalize.adj.method = c("laplacian", "row", "column", "none"),
  rwr.normalize.affinity = TRUE,
  rwr.threads = 2L,
  cells = NULL,
  features = NULL,
  verbose = TRUE,
  ...
)

\S4method{sc.rwr}{SingleCellExperiment}(
  data,
  gset.idx.list,
  gsvaExp.name = "gset1.rwr",
  min.sz = 10,
  max.sz = Inf,
  gene.occurrence.rate = 0.4,
  assay.type = "logcounts",
  knn.mca.consider.spcoord = TRUE,
  knn.used.reduction.dims = 30,
  knn.BPPARAM = SerialParam(),
  knn.graph.weighted = TRUE,
  knn.graph.directed = FALSE,
  knn.k.use = 500,
  rwr.restart = 0.7,
  rwr.normalize.adj.method = c("laplacian", "row", "column", "none"),
  rwr.normalize.affinity = TRUE,
  rwr.threads = 2L,
  cells = NULL,
  features = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object normalized and have the result of
\code{UMAP} or \code{TSNE}. Or a \linkS4class{SVPExperiment} object.}

\item{gset.idx.list}{gene set list contains the names.}

\item{gsvaExp.name}{a character the name of \code{gsvaExp} of result \code{SVP} object.}

\item{min.sz}{integer the minimum gene set number, default is 10, the number of gene sets
smaller than \code{min.sz} will be ignored.}

\item{max.sz}{integer the maximum gene set number, default is Inf, the number of gene sets
larger than \code{max.sz} will be ignored.}

\item{gene.occurrence.rate}{the occurrence proportion of the gene set in the input object,
default is 0.4.}

\item{assay.type}{which expressed data to be pulled to build KNN Graph, default is \code{logcounts}.}

\item{knn.mca.consider.spcoord}{logical whether consider the spatial coordinates to run MCA, default is TRUE.}

\item{knn.used.reduction.dims}{the top components of the reduction with \code{MCA} to be used to build KNN
Graph, default is 30.}

\item{knn.BPPARAM}{A BiocParallelParam object specifying whether the built of KNN should be parallelized
default is \code{SerialParam()}, meaning no parallel. You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)}
to parallel it, the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}.}

\item{knn.graph.weighted}{logical whether consider the distance of nodes in the Nearest Neighbors, default is TRUE.}

\item{knn.graph.directed}{logical whether consider the direction of the KNN Graph, default is FALSE.}

\item{knn.k.use}{numeric the number of the Nearest Neighbors nodes, default is 500.}

\item{rwr.restart}{default is 0.7.}

\item{rwr.normalize.adj.method}{character the method to normalize the adjacency matrix of the input graph,
default is \code{laplacian}.}

\item{rwr.normalize.affinity}{logical whether normalize the activity (affinity) result score using quantile normalisation,
default is TRUE.}

\item{rwr.threads}{the threads to run Random Walk With Restart (RWR), default is 2L.}

\item{cells}{Vector specifying the subset of cells to be used for the calculation of the activaty score or identification
of SV features. This can be a character vector of cell names, an integer vector of column indices or a logical vector,
default is NULL, meaning all cells to be used for the calculation of the activaty score or identification of SV features.}

\item{features}{Vector specifying the subset of features to be used for the calculation of the activaty score or identification
of SV features. This can be a character vector of features names, an integer vector of row indices or a logical vector,
default is NULL, meaning all features to be used for the calculation of the activaty score or identification of SV features.}

\item{verbose}{logical whether print the intermediate message when running the program, default is TRUE.}

\item{...}{additional parameters}
}
\description{
Calculate the activity of gene sets in spatial or single-cell data with Restart Walk with Restart
}
