% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svp.R
\name{svp}
\alias{svp}
\alias{svp,SingleCellExperiment,GeneSetCollection-method}
\alias{svp,SingleCellExperiment,GeneSetCollection}
\alias{svp,SingleCellExperiment,list-method}
\alias{svp,SingleCellExperiment,list}
\title{svp}
\usage{
svp(
  data,
  gset.idx.list,
  .assay,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.parallel.sz = 1L,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  ...
)

\S4method{svp}{SingleCellExperiment,GeneSetCollection}(
  data,
  gset.idx.list,
  .assay,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.parallel.sz = 1L,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  ...
)

\S4method{svp}{SingleCellExperiment,list}(
  data,
  gset.idx.list,
  .assay,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.parallel.sz = 1L,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  ...
)
}
\arguments{
\item{data}{object}

\item{gset.idx.list}{gene set list}

\item{.assay}{the assay name of \code{SingleCellExperiment}}

\item{gsva.method}{the method of gsva}

\item{gsva.kcdf}{the method of kcdf}

\item{gsva.abs.ranking}{logical whether rank the abs}

\item{gsva.min.sz}{integer the minimum gene set number}

\item{gsva.max.sz}{integer the maximum gene set number}

\item{gsva.parallel.sz}{the thread number}

\item{gsva.mx.diff}{logical}

\item{gsva.tau}{exponent defining the weight of the tail in the random walk}

\item{gsva.ssgsea.norm}{logical whether normalizing when \code{gsva.method = 'ssgsea'}}

\item{gsva.verbose}{logical whether print the message of process in gsva}

\item{gsva.BPPARAM}{An object of class \code{BiocParallelParam} specifiying parameters}

\item{...}{additional parameters}
}
\description{
svp
}
