% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVP-package.R
\docType{package}
\name{SVP-package}
\alias{SVP}
\alias{SVP-package}
\title{SVP: Predicting cell states and their variability in single-cell or spatial omics data}
\description{
SVP uses the distance between cells and cells, features and features, cells and features in the space of MCA to build nearest neighbor graph, then uses random walk with restart algorithm to calculate the activity score of gene sets (such as cell marker genes, kegg pathway, go ontology, gene modules, transcription factor or miRNA target sets, reactome pathway, ...), which is then further weighted using the hypergeometric test results from the original expression matrix. To detect the spatially or single cell variable gene sets or (other features) and the spatial colocalization between the features accurately, SVP provides some global and local spatial autocorrelation method to identify the spatial variable features. SVP is developed based on SingleCellExperiment class, which can be interoperable with the existing computing ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/SVP}
  \item Report bugs at \url{https://github.com/YuLab-SMU/SVP/issues}
}

}
\author{
\strong{Maintainer}: Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})

Authors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [contributor]
}

}
\keyword{internal}
