% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-wkde.R
\name{wkde}
\alias{wkde}
\alias{wkde,SingleCellExperiment-method}
\alias{wkde,SingleCellExperiment}
\alias{wkde,SVPExperiment-method}
\alias{wkde,SVPExperiment}
\title{Calculating the 2D Weighted Kernel Density Estimation}
\usage{
wkde(
  data,
  assay.type = "logcounts",
  reduction = c("UMAP", "TSNE"),
  grid.n = 100,
  adjust = 1,
  bandwidths = NULL,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{wkde}{SingleCellExperiment}(
  data,
  assay.type = "logcounts",
  reduction = c("UMAP", "TSNE"),
  grid.n = 100,
  adjust = 1,
  bandwidths = NULL,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{wkde}{SVPExperiment}(
  data,
  assay.type = "logcounts",
  reduction = c("UMAP", "TSNE"),
  grid.n = 100,
  adjust = 1,
  bandwidths = NULL,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object with contains \code{UMAP} or \code{TSNE},
or a \linkS4class{SpatialExperiment} object, or a \linkS4class{SVPExperiment} object with specified
\code{gsvaexp} argument.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{logcounts}.}

\item{reduction}{character used as spatial coordinates to detect SVG, default is \code{UMAP},
if \code{data} has \code{spatialCoords}, which will be used as spatial coordinates.}

\item{grid.n}{integer number of grid points in the two directions to estimate 2D weighted kernel density, default is 100.}

\item{adjust}{numeric to adjust the \code{bandwidths}, default is 1.0.}

\item{bandwidths}{vector a two length numeric vector represents the bandwidths for x and y directions, default is normal
reference bandwidth \code{\link[ks]{hpi}}, see also \code{\link[MASS]{bandwidth.nrd}}.}

\item{verbose}{logical whether print the intermediate message when running the program, default is TRUE.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{...}{additional parameters}
}
\value{
a \linkS4class{SVPExperiment} or \linkS4class{SingleCellExperiment}
}
\description{
Calculating the 2D Weighted Kernel Density Estimation
}
