% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lisaRes.R
\name{as_tbl_df}
\alias{as_tbl_df}
\title{convert the square matrix to long tidy table}
\usage{
as_tbl_df(x, listn = NULL)
}
\arguments{
\item{x}{list or matrix object, which is the output of \code{runGLOBALBV},
\code{fast_cor} or the matrix output of \code{cor}.}

\item{listn}{list object, which must have name, and the element must
from the row names of \code{x} or \code{x[[1]]} (when \code{x} is a list)
default is NULL.}
}
\value{
a long tidy table
}
\description{
This function is designed to convert the output of \code{runGLOBALBV},
\code{fast_cor} or the matrix output of \code{cor} to long tidy table.
}
\examples{
example(fast_cor, echo=FALSE)
x <- as_tbl_df(res)
head(x)
x2 <- as_tbl_df(res2)
head(x2)
}
