% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalBgSpatialKld}
\alias{CalBgSpatialKld}
\title{Compute Background 2D Kernel Density}
\usage{
CalBgSpatialKld(coords, axm, aym, h, indx, indy)
}
\arguments{
\item{coords}{coordinate matrix.}

\item{axm}{matrix the difference between the original point and grid points in x direction.}

\item{aym}{matrix the difference between the original point and grid points in y direction.}

\item{h}{The vector of bandwidths for x and y directions, defaults to normal reference bandwidth}

\item{indx}{the index of original point by mapping to the grid points in x direction.}

\item{indy}{the index of original point by mapping to the grid points in y direction.
(see MASS::bandwidth.nrd), A scalar value will be taken to apply to both directions (see ks::hpi).}
}
\description{
Compute Background 2D Kernel Density
}
