% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalBgSpatialKld}
\alias{CalBgSpatialKld}
\title{Compute Background 2D Kernel Density}
\usage{
CalBgSpatialKld(coords, n = 25L)
}
\arguments{
\item{coords}{coordinate matrix.}

\item{n}{the Number of grid points in each direction, default is 25.}
}
\description{
Compute Background 2D Kernel Density
}
