% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalSpatialKld}
\alias{CalSpatialKld}
\title{Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted and Statistical Test With Permutation for single weight vector.}
\usage{
CalSpatialKld(
  d,
  bgkld,
  axm,
  aym,
  h,
  indx,
  indy,
  random_times = 100L,
  seed = 1024
)
}
\arguments{
\item{d}{the weight vector (the expression of gene or score of pathway).}

\item{bgkld}{the kernel density of background (the result of CalBgSpatialKld).}

\item{axm}{matrix the difference between the original point and grid points in x direction.}

\item{aym}{matrix the difference between the original point and grid points in y direction.}

\item{h}{The vector of bandwidths for x and y directions, defaults to normal reference bandwidth
(see bandwidth.nrd), A scalar value will be taken to apply to both directions (see ks::hpi).}

\item{indx}{the index of original point by mapping to the grid points in x direction.}

\item{indy}{the index of original point by mapping to the grid points in y direction.}

\item{random_times}{the permutation numbers for each weight to test whether
it is significantly, default is 200.}

\item{seed}{The random seed to use to evaluate, default 123.}
}
\value{
a vector of input features about the statistical test value with 2D weighted kernel density
and Kullback–Leibler Divergence.
}
\description{
Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted and Statistical Test With Permutation for single weight vector.
}
