% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalSpatialKldCpp}
\alias{CalSpatialKldCpp}
\title{Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted And Statistical Test With Permutation.}
\usage{
CalSpatialKldCpp(coords, d, n = 25L, random_times = 999L)
}
\arguments{
\item{coords}{coordinate matrix.}

\item{d}{matrix (the expression of gene or score of pathway).}

\item{n}{the Number of grid points in each direction, default is 25.}

\item{random_times}{the permutation numbers for each weight to test whether
it is significantly, default is 999.}
}
\description{
Compute the Kullback–Leibler Divergence using 2D Kernel Density Estimation
With Weighted And Statistical Test With Permutation.
}
