% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPDist}
\alias{fastPDist}
\title{Obtain the pair distance of row between Ar and Br matrix}
\usage{
fastPDist(Ar, Br)
}
\arguments{
\item{Ar}{matrix which number of column should be equal to column number of Br.}

\item{Br}{matrix which number of column should be equal to column number of Ar.}
}
\description{
Obtain the pair distance of row between Ar and Br matrix
}
