% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findIntervalCpp}
\alias{findIntervalCpp}
\title{Find Interval Numbers or Indices in C++}
\usage{
findIntervalCpp(x, breaks)
}
\arguments{
\item{x}{numeric vector (orignial).}

\item{breaks}{numeric vector (new).}
}
\value{
the vector of length \code{length(x)} with values in \code{0:N}
this is like \code{findInterval()} of R base, but result of this is can be used
to the C++
}
\description{
Find Interval Numbers or Indices in C++
}
