% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svg.R
\name{kldSVG}
\alias{kldSVG}
\alias{kldSVG,SingleCellExperiment-method}
\alias{kldSVG,SingleCellExperiment}
\alias{kldSVG,SVPExperiment-method}
\alias{kldSVG,SVPExperiment}
\title{Detecting the spatially or single cell variable features with Kullback–Leibler
divergence of 2D Weighted Kernel Density}
\usage{
kldSVG(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "bonferroni",
  sv.BPPARAM = SerialParam(),
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{kldSVG}{SingleCellExperiment}(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "bonferroni",
  sv.BPPARAM = SerialParam(),
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{kldSVG}{SVPExperiment}(
  data,
  assay.type = "logcounts",
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = 100,
  sv.p.adjust.method = "bonferroni",
  sv.BPPARAM = SerialParam(),
  verbose = TRUE,
  random.seed = 1024,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object with contains \code{UMAP} or \code{TSNE},
or a \linkS4class{SpatialExperiment} object, or a \linkS4class{SVPExperiment} object with specified
\code{gsvaexp} argument.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{logcounts}.}

\item{sv.used.reduction}{character used as spatial coordinates to detect SVG, default is \code{UMAP},
if \code{data} has \code{spatialCoords}, which will be used as spatial coordinates.}

\item{sv.grid.n}{numeric number of grid points in the two directions to estimate 2D weighted kernel
density, default is 100.}

\item{sv.permutation}{numeric the number of permutation for each single feature to detect the
signicantly spatially or single cell variable features, default is 100.}

\item{sv.p.adjust.method}{character the method to adjust the pvalue of the result, default is \code{bonferroni}.}

\item{sv.BPPARAM}{A BiocParallelParam object specifying whether the identification of SV features should be parallelized
default is \code{SerialParam()}, meaning no parallel. You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)}
to parallel it, the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}. default is \code{SerialParam()}.}

\item{verbose}{logical whether print the intermediate message when running the program, default is TRUE.}

\item{random.seed}{numeric random seed number to repeatability, default is 1024.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{...}{additional parameters}
}
\value{
a \linkS4class{SVPExperiment} or a \linkS4class{SingleCellExperiment}, see details.
}
\description{
Detecting the spatially or single cell variable features with Kullback–Leibler
divergence of 2D Weighted Kernel Density
}
\details{
if input is a \linkS4class{SVPExperiment}, output will be also a \linkS4class{SVPExperiment}, the spatially variable gene sets
result is stored in \code{svDfs} of the specified \code{gsvaexp}, which is a \linkS4class{SingleCellExperiment}. If input is
a \linkS4class{SingleCellExperiment} (which is extracted from \linkS4class{SVPExperiment} using \code{gsvaExp()} funtion), output
will be also a \linkS4class{SingleCellExperiment}, the spatial variable gene sets result can be extracted using \code{svDf} function.
The result of \code{svDf} will return a matrix which has \code{sp.kld}, \code{boot.sp.kld.mean}, \code{boot.sp.kld.sd}, \code{pvalue},
\code{padj} and \code{rank}.
\itemize{
\item \code{sp.kld} which is logarithms of Kullback–Leibler divergence, larger value meaning the greater the difference from the
background distribution without spatial variability.
\item \code{boot.sp.kld.mean} which is mean of logarithms of Kullback–Leibler divergence based on the permutation of each features.
\item \code{boot.sp.kld.sd} which is standard deviation of logarithms of Kullback–Leibler divergence based on the permutation of
each features.
\item \code{pvalue} the pvalue is calculated using the real \code{sp.kld} and the permutation \code{boot.sp.kld.mean} and
\code{boot.sp.kld.sd} based on the normal distribution.
\item \code{padj} the adjusted pvalue based on the speficied \code{sv.p.adjust.method}, default is \code{BY}.
\item \code{rank} the order of significant spatial variable features based on \code{padj} and \code{sp.kld}.
}
}
\seealso{
\code{\link{sc.rwr}} to calculate the activity score of gene sets.
}
