% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{imgData}
\alias{imgData<-}
\alias{spatialCoordsNames}
\alias{spatialCoordsNames<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SpatialExperiment}{\code{\link[SpatialExperiment:SpatialExperiment-methods]{imgData}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{imgData<-}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoords}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoords<-}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoordsNames}}, \code{\link[SpatialExperiment:SpatialExperiment-methods]{spatialCoordsNames<-}}}
}}

