% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svg.R
\name{runDetectSVG}
\alias{runDetectSVG}
\alias{runDetectSVG,SingleCellExperiment-method}
\alias{runDetectSVG,SingleCellExperiment}
\alias{runDetectSVG,SVPExperiment-method}
\alias{runDetectSVG,SVPExperiment}
\title{Detecting the spatially or single cell variable features with Moran's I or Geary's C of
2D weighted kernel density estimation}
\usage{
runDetectSVG(
  data,
  assay.type = "logcounts",
  method = c("moransi", "gearysc"),
  weight = NULL,
  weight.method = c("knn", "tri2nb", "none"),
  sv.runWKDE = FALSE,
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = NULL,
  sv.p.adjust.method = "BH",
  random.seed = 1024,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{runDetectSVG}{SingleCellExperiment}(
  data,
  assay.type = "logcounts",
  method = c("moransi", "gearysc"),
  weight = NULL,
  weight.method = c("knn", "tri2nb", "none"),
  sv.runWKDE = FALSE,
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = NULL,
  sv.p.adjust.method = "BH",
  random.seed = 1024,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)

\S4method{runDetectSVG}{SVPExperiment}(
  data,
  assay.type = "logcounts",
  method = c("moransi", "gearysc"),
  weight = NULL,
  weight.method = c("knn", "tri2nb", "none"),
  sv.runWKDE = FALSE,
  sv.used.reduction = c("UMAP", "TSNE"),
  sv.grid.n = 100,
  sv.permutation = NULL,
  sv.p.adjust.method = "BH",
  random.seed = 1024,
  verbose = TRUE,
  gsvaexp = NULL,
  gsvaexp.assay.type = NULL,
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object with contains \code{UMAP} or \code{TSNE},
or a \linkS4class{SpatialExperiment} object, or a \linkS4class{SVPExperiment} object with specified
\code{gsvaexp} argument.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{logcounts}.}

\item{method}{character one of \code{'moransi'} and \code{"gearysc"}, default is \code{'moransi'}.}

\item{weight}{object, which can be \code{nb}, \code{listw} or \code{Graph} object, default is NULL.}

\item{weight.method}{character the method to build the spatial neighbours weights, default
is \code{knn} (k nearest neighbours). Other method, which requires coord matrix as input and returns
\code{nb}, \code{listw} or \code{Graph} object, also is avaiable, such as \code{'tri2nb'}, \code{"knearneigh"},
\code{'dnearneigh'}, \code{"gabrielneigh"}, \code{"relativeneigh"}, which are from \code{spdep} package.
default is \code{knn}, if it is \code{"none"}, meaning the distance weight of each spot is used to
the weight.}

\item{sv.runWKDE}{logical whether perform the 2D weighted kernel density estimation firstly,
default is FALSE, it is experimental parameter.}

\item{sv.used.reduction}{character used as spatial coordinates to detect SVG, default is \code{UMAP},
if \code{data} has \code{spatialCoords}, which will be used as spatial coordinates.}

\item{sv.grid.n}{numeric number of grid points in the two directions to estimate 2D weighted kernel
density, default is 100.}

\item{sv.permutation}{integer the number to permutation test for the calculation of Moran's I, default
is NULL.}

\item{sv.p.adjust.method}{character the method to adjust the pvalue of the result, default is \code{BY}.}

\item{random.seed}{numeric random seed number to repeatability, default is 1024.}

\item{verbose}{logical whether print the intermediate message when running the program, default is TRUE.}

\item{gsvaexp}{which gene set variation experiment will be pulled to run, this only work when \code{data} is a
\linkS4class{SVPExperiment}, default is NULL.}

\item{gsvaexp.assay.type}{which assay data in the specified \code{gsvaexp} will be used to run, default is NULL.}

\item{...}{additional parameters}
}
\value{
a \linkS4class{SVPExperiment} or a \linkS4class{SingleCellExperiment}, see details.
}
\description{
To resolve the sparsity of single cell or spatial omics data, we use kernel function smoothing cell
density weighted by the gene expression in a low-dimensional space or physical space. This method had
reported that it can better represent the gene expression, it can also recover the signal from cells that
are more likely to express a gene based on their neighbouring cells (first reference). Next, we use
Moran's I or Geary's C to detect the signal genes in a low-dimensional space (\code{UMAP} or \code{TSNE}
for single cell omics data) or a physical space (for spatial omics data).
}
\examples{
# This example dataset is extracted from the
# result of runSGSA with gsvaExp(svpe).
data(hpda_spe_cell_dec)

hpda_spe_cell_dec <-
    hpda_spe_cell_dec |>
    runDetectSVG(
      assay.type = 'affi.score',
      method = 'moransi'
    )
# The result also is saved in the svDfs in the SVPExample object
# which can be extrated with svDf 
svDfs(hpda_spe_cell_dec)

hpda_spe_cell_dec |> svDf("sv.moransi") |> data.frame() |> dplyr::arrange(rank)

hpda_spe_cell_dec <-
    hpda_spe_cell_dec |>
    runDetectSVG(assay.type ='affi.score', method = 'gearysc')

svDfs(hpda_spe_cell_dec)

hpda_spe_cell_dec |> svDf("sv.gearysc") |> data.frame() |> dplyr::arrange(rank)
}
\references{
\enumerate{
\item Jose Alquicira-Hernandez, Joseph E Powell, Nebulosa recovers single-cell gene expression signals by kernel density estimation.
Bioinformatics, 37, 2485–2487(2021), https://doi.org/10.1093/bioinformatics/btab003.
}
}
\author{
Shuangbin Xu
}
