% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lisa.R
\name{runLISA}
\alias{runLISA}
\alias{runLISA,SingleCellExperiment-method}
\alias{runLISA,SingleCellExperiment}
\title{Local indicators of spatial association analysis}
\usage{
runLISA(
  data,
  features,
  assay.type = "logcounts",
  method = c("localG", "localmoran"),
  weight = NULL,
  weight.method = c("knn", "tri2nb"),
  reduction.used = NULL,
  cells = NULL,
  action = c("get", "add", "only"),
  alternative = "two.sided",
  BPPARAM = SerialParam(),
  ...
)

\S4method{runLISA}{SingleCellExperiment}(
  data,
  features,
  assay.type = "logcounts",
  method = c("localG", "localmoran"),
  weight = NULL,
  weight.method = c("knn", "tri2nb"),
  reduction.used = NULL,
  cells = NULL,
  action = c("get", "add", "only"),
  alternative = "two.sided",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{data}{a \linkS4class{SingleCellExperiment} object with contains \code{UMAP} or \code{TSNE},
or a \linkS4class{SpatialExperiment} object, or a \linkS4class{SVPExperiment} object with specified
\code{gsvaexp} argument.}

\item{features}{the feature name or index of data object, which are required.}

\item{assay.type}{which expressed data to be pulled to run, default is \code{logcounts}.}

\item{method}{character one of \code{'localG'}, \code{"localmoran"}, default is \code{'localG'}.}

\item{weight}{object, which can be \code{nb}, \code{listw} or \code{Graph} object, default is NULL.}

\item{weight.method}{character the method to build the spatial neighbours weights, default
is \code{knn} (k nearest neighbours). Other method, which requires coord matrix as input and returns
\code{nb}, \code{listw} or \code{Graph} object, also is avaiable, such as \code{'tri2nb'}, \code{"knearneigh"},
\code{'dnearneigh'}, \code{"gabrielneigh"}, \code{"relativeneigh"}, which are from \code{spdep} package.
default is \code{knn}, if it is \code{"none"}, meaning the distance weight of each spot is used to
the weight.}

\item{reduction.used}{character used as spatial coordinates to calculate the neighbours weights,
default is \code{UMAP}, if \code{data} has \code{spatialCoords}, which will be used as spatial coordinates.}

\item{cells}{the cell name or index of data object, default is NULL.}

\item{action}{character, which control the type of return result, default is \code{get}, which will return
a \linkS4class{SimpleList}.}

\item{alternative}{a character string specifying the alternative hypothesis, default is \code{two.sided}.}

\item{BPPARAM}{A BiocParallelParam object specifying whether perform the analysis parallelly using
\code{BiocParallel} default is \code{SerialParam()}, meaning no parallel.
You can use \code{BiocParallel::MulticoreParam(workers=4, progressbar=T)} to parallel it,
the \code{workers} of \code{MulticoreParam} is the number of cores used, see also
\code{\link[BiocParallel]{MulticoreParam}}. default is \code{SerialParam()}..}

\item{...}{additional parameters the parameters which are from the weight.method function.}
}
\value{
if \code{action = 'get'} (in default), the SimpleList object (like list object) will be return,
if \code{action = 'only'}, the data.frame will be return. if \code{action = 'add'}, the result of LISA is
stored in the \code{localResults} column of \code{int_colData} (internal column metadata). You can use
\code{localResults()} function of \code{SpatialFeatureExperiment} package to extract it.
}
\description{
This function use the local indicators of spatial association (LISA) to identify the hotspot
in the spatial space.
}
\references{
\enumerate{
\item Bivand, R.S., Wong, D.W.S. Comparing implementations of global and local indicators of spatial association. TEST 27,
716–748 (2018). https://doi.org/10.1007/s11749-018-0599-x
}
}
