% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-accessor.R, R/methods-show.R
\docType{methods}
\name{SVP-accessors}
\alias{SVP-accessors}
\alias{spatialCoords,SVPExperiment-method}
\alias{spatialCoordsNames,SVPExperiment-method}
\alias{imgData,SVPExperiment-method}
\alias{show,SVPExperiment-method}
\title{Some accessor funtions to get the internal slots of SVPExperiment}
\usage{
\S4method{spatialCoords}{SVPExperiment}(x)

\S4method{spatialCoordsNames}{SVPExperiment}(x)

\S4method{imgData}{SVPExperiment}(x)

\S4method{show}{SVPExperiment}(object)
}
\arguments{
\item{x}{a \linkS4class{SVPExperiment} class.}

\item{object}{a \linkS4class{SVPExperiment} class.}
}
\value{
matrix or character or print the information of object.
}
\description{
Some accessor funtions to get the internal slots of SVPExperiment
}
\examples{
library(SpatialExperiment) |> suppressPackageStartupMessages()
library(DropletUtils) |> suppressPackageStartupMessages()
example(read10xVisium, echo = FALSE)
svpe <- as(spe, 'SVPExperiment')
svpe
spatialCoords(svpe) |> head()
}
