% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-svp.R
\name{svp}
\alias{svp}
\alias{svp,SingleCellExperiment,GeneSetCollection-method}
\alias{svp,SingleCellExperiment,GeneSetCollection}
\alias{svp,SingleCellExperiment,list-method}
\alias{svp,SingleCellExperiment,list}
\alias{svp,SpatialExperiment,GeneSetCollection-method}
\alias{svp,SpatialExperiment,GeneSetCollection}
\alias{svp,SpatialExperiment,list-method}
\alias{svp,SpatialExperiment,list}
\title{svp}
\usage{
svp(
  data,
  gset.idx.list,
  assayName,
  threads = 1L,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  sv.X = NULL,
  sv.n_neighbors = 10,
  sv.order = "AMMD",
  sv.verbose = FALSE,
  sv.cor.method = c("pearson", "spearman", "kendall"),
  sv.cor.padj.method = "fdr",
  sv.cor.image.beta = 2,
  ...
)

\S4method{svp}{SingleCellExperiment,GeneSetCollection}(
  data,
  gset.idx.list,
  assayName,
  threads = 1L,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  sv.X = NULL,
  sv.n_neighbors = 10,
  sv.order = "AMMD",
  sv.verbose = FALSE,
  sv.cor.method = c("pearson", "spearman", "kendall"),
  sv.cor.padj.method = "fdr",
  sv.cor.image.beta = 2,
  ...
)

\S4method{svp}{SingleCellExperiment,list}(
  data,
  gset.idx.list,
  assayName,
  threads = 1L,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  sv.X = NULL,
  sv.n_neighbors = 10,
  sv.order = "AMMD",
  sv.verbose = FALSE,
  sv.cor.method = c("pearson", "spearman", "kendall"),
  sv.cor.padj.method = "fdr",
  sv.cor.image.beta = 2,
  ...
)

\S4method{svp}{SpatialExperiment,GeneSetCollection}(
  data,
  gset.idx.list,
  assayName,
  threads = 1L,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  sv.X = NULL,
  sv.n_neighbors = 10,
  sv.order = "AMMD",
  sv.verbose = FALSE,
  sv.cor.method = c("pearson", "spearman", "kendall"),
  sv.cor.padj.method = "fdr",
  sv.cor.image.beta = 2,
  ...
)

\S4method{svp}{SpatialExperiment,list}(
  data,
  gset.idx.list,
  assayName,
  threads = 1L,
  gsva.method = c("gsva", "ssgsea", "zscore", "plage"),
  gsva.kcdf = c("Gaussian", "Poisson", "none"),
  gsva.abs.ranking = FALSE,
  gsva.min.sz = 1,
  gsva.max.sz = Inf,
  gsva.mx.diff = TRUE,
  gsva.tau = switch(gsva.method, gsva = 1, ssgsea = 0.25, NA),
  gsva.ssgsea.norm = TRUE,
  gsva.verbose = TRUE,
  gsva.BPPARAM = SerialParam(progressbar = gsva.verbose),
  sv.X = NULL,
  sv.n_neighbors = 10,
  sv.order = "AMMD",
  sv.verbose = FALSE,
  sv.cor.method = c("pearson", "spearman", "kendall"),
  sv.cor.padj.method = "fdr",
  sv.cor.image.beta = 2,
  ...
)
}
\arguments{
\item{data}{object}

\item{gset.idx.list}{gene set list}

\item{assayName}{the assay name of \code{SingleCellExperiment}}

\item{threads}{the thread number, default is 1.}

\item{gsva.method}{the method of gsva}

\item{gsva.kcdf}{the method of kcdf}

\item{gsva.abs.ranking}{logical whether rank the abs}

\item{gsva.min.sz}{integer the minimum gene set number}

\item{gsva.max.sz}{integer the maximum gene set number}

\item{gsva.mx.diff}{logical}

\item{gsva.tau}{exponent defining the weight of the tail in the random walk}

\item{gsva.ssgsea.norm}{logical whether normalizing when \code{gsva.method = 'ssgsea'}}

\item{gsva.verbose}{logical whether print the message of process in gsva}

\item{gsva.BPPARAM}{An object of class \code{BiocParallelParam} specifiying parameters}

\item{sv.X}{matrix default is NULL.}

\item{sv.n_neighbors}{integer default is 10.}

\item{sv.order}{character default is \code{AMMD}.}

\item{sv.verbose}{logical}

\item{sv.cor.method}{character, the method to calculate the correlation of between
the features and the color value for the RGB channels of histology image.}

\item{sv.cor.padj.method}{character, the method of p.adjust for the correlation test}

\item{sv.cor.image.beta}{integer, the square pixels to calculate the color value for
the RGB channels of histology image, default 2, meaning 2 x 2 pixels square.}

\item{...}{additional parameters}
}
\description{
svp
}
