% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVP-package.R
\docType{package}
\name{SVP-package}
\alias{SVP}
\alias{SVP-package}
\title{SVP: Efficient analysis of 'gene set' activity in spatial or single-cell data}
\description{
SVP uses the distance between cells and cells, features and features, cells and features in the space of MCA to build nearest neighbor graph, then uses random walk with restart algorithm to calculate the activity score of gene sets (such as kegg pathway, signatures, go term, gene modules, transcription factor, ...). In addition, to detect the spatially or single cell variable gene sets or (other features) accurately, SVP uses the 2d weighted kernel density estimation to process the score of gene sets (or expression of genes) and uses Kullback–Leibler divergence to identify the signal features based on permutation test. SVP is developed based on SingleCellExperiment class, which can be interoperable with the existing computing ecosystem.
}
\author{
\strong{Maintainer}: Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [contributor]
}

}
\keyword{internal}
